server/ml/requirements.txt
flask>=3.0
torch>=2.6
torchvision>=0.21
pillow>=10.4
opencv-python>=4.10
numpy>=1.26
